pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pysoar"
version = "1.0.0"
description = "Security Orchestration, Automation and Response Platform"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "PySOAR Team", email = "team@pysoar.local"}
]
keywords = ["security", "orchestration", "automation", "response", "siem", "soc"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "aiosqlite>=0.19.0",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "flower>=2.0.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.5.3",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",
    "orjson>=3.9.0",
    "python-dateutil>=2.8.2",
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",
    "aiosmtplib>=3.0.1",
    "email-validator>=2.1.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
    "python-dotenv>=1.0.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "tenacity>=8.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "faker>=22.2.0",
    "mypy>=1.8.0",
    "black>=24.1.0",
    "isort>=5.13.2",
    "ruff>=0.3.0",
    "bandit>=1.7.5",
    "safety>=2.3.5",
    "pip-audit>=2.6.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"
select = [
    "E",
    "F",
    "W",
    "I",
    "N",
    "B",
    "A",
    "COM",
    "C4",
    "ARG",
    "RUF",
]
ignore = [
    "E501",
    "B008",
    "A003",
]

[tool.ruff.isort]
known-first-party = ["src"]

[tool.black]
line-length = 120
target-version = ["py310", "py311"]
extend-exclude = '''
/(
  | \.git
  | \.venv
  | venv
  | __pycache__
  | \.pytest_cache
  | dist
  | build
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]
known_first_party = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
plugins = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__init__.py",
    "*/conftest.py",
]
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if 0:",
    "if False:",
]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"
